requirements.txt
Django~=1.11.29
djangorestframework~=3.9.4
bcrypt~=3.1.7
requests~=2.24.0